MGTK: Process mouse moves in same tick as button state changes
Investigating some weirdness with Apple //jse with touch events, it
became apparent that when the mouse state updated with both position
and button state, a button_down event would be dispatched with the
*previous* position before an event with the new position. This seems
to be due to weird ordering in MoveCursor.

Reordering so that the mouse is polled first addresses this. It also
renders the change in 7a9834d2 unnecessary and clarifies the
previously poorly named `set_pos_params` is really the cursor
position, so rename that variable and undo that previous change, and
also update the "did the mouse move?" logic for the menu event loop to
use the cursor position instead of the mouse position.

Note that this doesn't seem to fix the weirdness seen in //jse but
seems like a good change anyway. In theory this should mildly improve
responsiveness.