IconTK: Move "layout fixed?" from icon flag to DragHighlighted param
When selection in list views (e.g. View > by Name) was introduced, a
flag was added to icons to signal "this icon can't be moved". This
worked fine; during a drag operation it would be used to disallow
moving icons within a window.

To save one bit per icon (useful in the future?), use the window's
view style as a flag passed into the IconTK call instead. This is
simpler anyway.