DeskTop: Prevent crash launching programs after previewing images.
The INVOKER_PREFIX buffer was re-used by image preview and not
cleared; this caused the invoker to think it should launch the next
program (e.g. system file, binary file) with that path as an
interpreter. Boom!

Reset the path at the start of the launch sequence, as it will be
filled in by the rest of the code with an interpreter if needed. This
was not the only bit of code that re-used that buffer so there were
probably other repros.

Fixes #765