DeskTop: Partially restore /RAM before invoking Disk Copy
Disk Copy dutifully saves/restores /RAM on entry/exit, but DeskTop
wasn't restoring the driver address it before invoking Disk Copy, so
DeskTop -> Disk Copy -> DeskTop -> BASIC.SYSTEM would leave /RAM in a
bad state.

The usual ReconnectRAM proc isn't used because the copy of MGTK in
aux and the DHR screen are retained across DeskTop -> Disk Copy. It's
weird but it works. So just use a subset with lots of documentation.

Fixes #766