Invoker: Populate $280 with full prefix of target file
I'd misinterpreted (I guess) @JohnMBrooks description of how Bitsy Bye
launches BASIS.SYSTEM as having $380 contain the volume path e.g.
"/VOL". From testing with Bitsy Bye, It actually contains the full
path of the target file e.g. "/VOL/MEDIA". This is redundant with the
PREFIX which is also set to that path, but it's better to be
compatible.