MGTK: Add FlashMenuBar command
This can be used to invert the menu bar, e.g. for silent alerts.

The Sounds control panel DA should be switched over to use this for
the "Silent" sound proc, but that'll require more work because the
banking for sound procs is undefined - the DA runs them with aux
banked in, DeskTop runs them with main banked in. This will need to be
made consistent to enable MGTK use from sound procs.