Sounds: Implement "Silent" via MGTK::FlashMenuBar
This ended up being a bit convoluted. Sound procs were not set up to
call MGTK since banking is inconsistent, and that seems like a bad
dependency to take anyway. So instead a sound proc starting with 0
(BRK) signals to the Bell proc in `lib/bell.s` code to flash the menu
bar instead. The Sounds DA is also simplified to delegate playing the
selected sound to DeskTop's copy of the Bell proc rather than
duplicating some of the code.

That's fine, but now the Bell proc needs to be able to make MGTK
calls. Fortunately just a minor tweak to Disk Copy was needed.

But now that Bell is making an MGTK call, it can change graphics
state. While the active GrafPort is saved/restored around the call,
since we run in optimized mode where the GrafPort is not updated from
the zero page, changes like calling SetPortBits can be lost. So after
lib/alert.s calls MGTK::SetPortBits, force a write back to the
GrafPort via MGTK::GetPort so that a subsequent Bell call doesn't
confuse things. Wheee!