Strings: Replace 'Loading' with 'Starting' and improve l10n
The "Loading XXX..." string template was used by DeskTop and Selector.
Make it a little friendlier by saying "Starting XXX..." instead. Also
apply to the non-English versions.

And then that brings up the fact that a template doesn't work for
French where it should be "Démarrage d'Apple II DeskTop..." but
"Démarrage des Raccourcis...". So stop using a template and embed the
product/module name in the localization string tables. This makes it
more fragile (easier to forget to update a string if a name changes)
but more correct.

In the future we could add a "assert string A contains string B" but
you can't do that in ca65 so I'm not bothering for now.