File Dialog: Improve API, simplify Add/Edit a Shortcut logic
Add/Edit a Shortcut has special initialization of the file dialog
since it can start off with a path selected. This meant it called many
"internal" procs for the file dialog.

Add a special entry point instead, so that the init logic is
simplified. Also, rename 'DeviceOnLine' to better explain what it is
actually doing.