IconTK: Remove unnecessary drag-outline redrawing.
When dragging icons, the logic was to xdraw (erase) the outlines when
a move happened, update the poly coords, and xdraw (draw) the
outlines. So far so good. But when an icon was highlighted (or
unhighlighted) the outlines were toggled around those calls, so two
extra xdraws and a little more jank at 1MHz.

Also, for some reason, detecting an icon to highlight was not done
when the mouse moved, but in the next tick, adding a little more
delay. Reorder the tests to improve that, although more streamlining
can be done.

This turned up a latent bug where the right GrafPort for drawing the
icon wasn't selected, so that needs to be added in.

No functional changes.