DeskTop: Prevent spurrious clicks following a double-click
When trying out real hardware (Apple IIc Plus) I was observing that
after a double-click to open a folder, another click would often come
through, causing an unexpected selection or window activation. The
code handling the double-clicks looks fine and it seemed to be maybe
timing related.

My theory is that with MGTK, sometimes a PeekEvent followed by a
GetEvent doesn't necessarily consume the peeked event, in this case
leaving a button down event in the queue. Instead of the GetEvent when
a double click was seen (via PeekEvent), perform a FlushEvents call.
This seems to make the problem go away, so I'm gonna call it "fixed"
but my confidence is only moderate.