IconTK: Simplify icon clipping logic slightly. Saves 12 bytes.
In the preamble to the complicated icon clipping logic, the left, top,
and right bounds of the overlapping window are decremented by a pixel
before diving in to determine how to adjust the clipping rect. After
staring at the math for a bit, it turns out that for the left/top
cases it is because of how "greater than" test is done - the test was
implemented as a "greater than or equals" (c/o BMI, skipping if "less
than"). If the test is inverted then it can properly be a "greater
than" (c/o BPL, skipping if "greater than or equal") and the decrement
can be skipped.

The adjustment for the right side is more subtle, so leaving that
alone for now.