DeskTop: Remove 12 volume limit from Format/Erase picker
The "Option Picker" logic used in Format/Erase (and DeskTop and
Selector's shortcut pickers) required a power-of-two number of rows
and dialog size constraints limited the Format/Erase dialog to 3
columns, so a maximum of 4*3=12 volumes could be shown. DeskTop
supports handling the ProDOS-8 maximum of 14 volumes, so some would be
elided.

Fix the Option Picker logic to handle arbitrary number of rows, and
tweak the dialog layout to allow up to 15 volumes to be shown in the
picker (although 14 is the max).