File Dialog: Simplify clients by handling LineEdit activation
If extra controls are active, don't make the caller interact with the
LineEdit - just do it when opening the window.

Also, since the `extra_controls_flag` is initialized to 0 in `Init`
there's no need for callers to reset it before calling `OpenWindow` if
they don't need controls.

No functional changes.