Dialogs: Don't need to FindWindow when testing for I-beam cursor
On mouse move, dialogs with LineEdit controls need to test if the
cursor is over the control and update the cursor shape appropriately.
The File Dialog and DeskTop's Rename (etc) dialog first did a
FindWindow to see if the cursor was over the window - that's not
necessary since the dialog is modal! Just delete the code, everything
is faster and smaller (-27 bytes each).

No functional changes.