DeskTop: Perform fewer GrafPort shenanigans during window updates
MGTK sends an "update" event to windows when they have been revealed
e.g. by a window on top of them moving or closing. This provides
a clipped GrafPort allowing for minimal updates.

DeskTop would take this port and transfer it into the Winfo's port
then delegate to the normal window entry drawing routines which
would select this port - extra work for little benefit. What does
need to be done is port offsets to account for the header, but
that can be simplified.

This moves some logic to the LC but probably saves ~70 bytes and
several calls in main.