formfast: Added Support for Titan Accelerator IIe
I happen to own this card, and it was missing from the DA, so this small
patch adds support for it.
I don't think it can interfere with anything, as it's using a write on
an otherwise read-only address ($c086)

Thank you for the amazing work on this great project!

Signed-off-by: Michel Pollet <buserror@gmail.com>