sound DA: Add a cancel button
I've been debugging my own IIe emulator lately, and I used the sound DA
a lot to test it... However it was always nagging me that this modal
dialog lacks a... cancel button!

So I dived in, I have to say I'm very fresh to the codebase, but I tried
to respect the 'style' and behaviours.

Signed-off-by: Michel Pollet <buserror@gmail.com>