Add GetNextEvent lib that synthesizes mouse moved events
This replaces looking for `no_event` and calling `CheckMouseMoved`
with a wrapper that synthesizes a `kEventKindMouseMoved`. It's a
little bit of extra overhead so callers in DeskTop don't need to all
update but it should be harmless. The name is inspired by a similar
helper in the old Macintosh toolbox.

Maybe a "WaitNextEvent" - which blocks until a useful event happens -
can be added in the future, but the lack of system timers makes things
trickier. I tried adding itthis but the ergonomics weren't obviously
better - it was possible to automagically call SystemTask every tick,
but if there was a LineEdit requiring LETK::Idle or other locally
periodic task then it couldn't be used. I further tried an "idle" or
"blink caret" synthetic event type but that ended up being fairly
intrusive, requiring knowledge of system settings.