IIc: reset 80 col firmware mode byte on startup
If a previous app smashed the screen holes on the IIc/IIc+, the
loading progress bar would fail to display mousetext, showing
WVWVWV... as the background instead.

Reset the "mode byte" of the 80 column firmware on the IIc, in the
$7FB screen hole reserved for slot 3. If $FF, the firmware detects it
is inactive and reinitializes properly. Any other value appears to
cause it to not handle MouseText output correctly.

(External programs should really be careful not to mess with screen
holes, but since we reset INVFLG to $FF anyway this is a trivial
addition.)