MGTK: Remove ZP stash for keyboard window moves/grows. Saves 80 bytes
It's unclear why, but the logic MGTK used for handling keyboard-driven
window moves/grows stashed the upper half of the zero page. I can't
see anything it's doing that makes this necessary, and that's a lot of
memory for no clear benefit. So remove that code and the buffer.

Given this, chop those 80 bytes off the size reserved for MGTK and
shuffle things down, saving a disk block here and there.

Also, update the READMEs listing segments which had grown a bit stale.
Also, remove a few unused symbols and scope some definitions.