Selector: Restore 40-col text mode when quick-booting slot
When Selector is launching you can hold Apple-1...7 to boot from a
slot. This didn't exit 80-column mode so the display was wonky if not
reset by the OS.

Additionally, the logic to clear the screen via HOME seemed to not
properly clear the auxmem half of the 80 column screen. Debugging
showed that 80STORE needed to be set (i.e. PAGE1/2 control banking) in
the 80-column firmware, which is unexpected - you'd expect the
firmware to do this!

Also, use case-folding rather than testing against upper- and
lower-case versions of characters, which is what started this whole
thing.