Enable 80STORE before interacting w/ 80col firmware
The 80-column firmware expects 80STORE to be active when it is active,
so calls like HOME will not correctly clear the aux bank if not set.
Since DeskTop etc. normally run with 80STORE off or in indeterminate
state, explicitly set it in paths that interact with the firmware,
e.g. clearing the screen when invoking another app or returning to
ProDOS.