BTK: Pass button record in calls, rather than params struct
For hobgoblin-ish consistency, the button toolkit defined a button
record (window, label, state, rect, etc) but calls to toolkit passed a
separate params block. The block contained the address of the button
record, plus room for other parameters. The only other parameter was
"update", used in BTK::Draw for when responding to an update event.

This added 3 bytes of overhead per button, and cluttered the usage
since there would be a `button_rec` and a `button_params` definition
for each.

Simplify and just pass the button record itself; introduce a separate
BTK::Update call which is like BTK::Draw but for use in update
processing i.e. it doesn't set the port. Also simplifies BTK itself.

Saves 79 bytes in Desktop (split among 2 segments) and more for DAs,
Selector and Disk Copy modules.