IconTK: Correctly bound dimming rendering when painting icons.
The change in a059ecca optimized OffsetPortAndIcon to not bother
updating the bottom/right bounds of the icon rect as it was not used
when clipping. But it was used when drawing the dimming effect for
open folders. This caused a temporary visual glitch during the
drawing, with a flash of the dimming extending well beyond the icon.

(The shading effect is implemented by first drawing a dimmed
rectangle, then drawing the icon over it, then XOR-drawing the dimming
rectangle again, thus leaving the dimming only over the icon. So the
effect extending far beyond the icon was only temporary, but notable.)

Restore the elided lines, while keeping the clipping optimizations
from that change.