MGTK: Expose Save/RestoreScreenRect. Saves 291 bytes per lib/alert
The alert code saves/restores the background to make tearing down
alerts nearly instant, and without requiring the underlying windows to
perform updates. This logic is very similar to the code in MGTK to do
"savebehind" for the menus, but implemented separately.

Dedupe the logic by exposing the guts of MGTK's menu savebehind as two
new MGTK API entry points that take a Rect. This makes MGTK slightly
larger, but the simplifications to lib/alert net out to well over a
entire page of memory.