New DA: Change Type
Shows the type and auxtype of selected file(s) and lets you edit
them. The fields will show blank if the types/auxtypes differ
among the selected files.

Note that the type is shown (and must be entered) using hex notation;
friendly names like SYS and TXT are not supported.

Resolves #149