DeskTop: Allow selection to remain in inactive windows.
Since vintage versions, a distinction has been made in the code
between the active (i.e. top-most) window (if any) and the "selected"
window i.e. the window containing the selected icons (or the desktop).
This came with limitations, as only the active window could be drawn
into, so many, many fixes have been made over time to limit selection
to the active window or desktop, clearing selection in other cases.

But since recent changes to IconTK allow painting arbitrarily clipped
icons in windows, this restriction can be relaxed. Various places that
would clear selection can now allow selection to be maintained, and a
few places where volume icon behavior was special-cased can be
relaxed, e.g.:

* If you click on a window's content area the owning icon folder icon
  can be selected even if it's in a background window.

* If you select multiple folder icons then File > Open, the icons can
  stay selected.

This revises how several issues were treated, including #364, #415,
(e.g. System 7.5.3) but not identical. For example, clicking on the
non-content parts of a window (title bar, scroll bars, header) will
activate it but will not clear selection and select the owning icon,
or the volume icon if not present.

Testers should watch out for any confusion that's crept in between the
selected and active window, e.g. when an operation on selection (in an
inactive window) causes the active window to be refreshed instead.