DeskTop: Fix formatting of inserted smartport disks.
Auditing the handling of DEVLST entries and unit number masking
revealed a case where an unmasked entry was passed to the format
dialog, which would result in a bogus search for the unit index.

Mask it! Also, remove some dead code and mask one more call site that
works without masking but really should mask.