DeskTop: Make icon -> small icon mapping table driven
Only "generic" and "folder" small icons are currently supported, so
the code did a simple "if icon type is folder, use small folder"
branch. This is fine but harder to extend.

Replace this with a table lookup, which takes more bytes but it's for
a good cause, in case we want to add more small icon types in the
future.