DeskTop: Add system folder icon for auxtype $8000
You can modify the otherwise unused* auxtype of a folder (DIR) to be
$8000 using the new Change Type DA or an command line utility. If so,
it will now get a special icon.

If you install A2D onto a hard disk in a subdirectory you could use
this approach to give the folder a different look.

* = I asked around; ProLine uses the bits for Unix-style permissions.
The risk of collisions seems low.