MGTK: Improve window title bar appearance
Rather than filling the entire title bar with stripes and making the
"go away" box full height, add a pixel of spacing above/below. This is
closer to the classic Macintosh and IIgs style. I was skeptical at
first but at least subjectively it does make the the title bars less
visually noisy.