Add MulDiv routine
This does a 16x16-bit multiply then divides the 32-bit intermediate
result by a 16-bit divisor. This approach is commonly used for scaling
e.g. R = A * N/D to scale A by (N/D), when doing scroll bar
calculations.

* DeskTop's scroll routines are updated to use this.
* Show Text File DA's scroll routines are updated to use this.
* Map DA's coordinate<->pixel routines are updated to use this.
* Remove the separate 32-bit multiply/divide routines.

The routine could probably be further optimized for space. It would
also be nice to include this in MGTK and dedupe with the routine it
uses.