SmartPort: Add support for ProDOS 2.4.3 mapping tables
The logic for determining SmartPort dispatch address and unit numbers
given a ProDOS unit number does... unfortunate things. It looks up the
ProDOS device address for a ProDOS unit number and grovels around
inside ProDOS for the internal translation tables which are not
otherwise exposed. This is unfortunately version specific.

The ProDOS 2.4.3 "dev" release moves the driver address, so the lookup
fails. This is harmless for DeskTop (which uses this lib/), in the
sense that the logic returns "failure" when the dispatch address can't
be determined, which means things like the name lookup fail safely,
eject requests are ignored, etc. Not catastrophic, at least.

To address this, optimize the code so it takes less space by relying
on invariants with the remapping driver, and handle the new base
address. Slightly more bytes than before and now handles the 2.4.3
driver address. Some day we will do better, I promise.