MGTK: Ensure TrackThumb reports no move if thumb doesn't move
When TrackThumb completes it calculates a new thumbpos, and tests
if the pos changed. If it didn't change, it shouldn't report a move.
But this calculation is lossy and can end up reporting a move when
the thumb wasn't moved at all, causing the thumb to jump unexpectedly.

Verify that the thumb really did move before signaling a move.

Fixes #778