Launcher: Prevent crash or hang on unsupported system
The DESKTOP.SYSTEM launcher tests early on for a IIe-or-later system
and 128K of RAM using ProDOS's MACHID flags. If unsupported, it QUITs
back to ProDOS. Unfortunately, the code has evolved so that it does
that after loading settings into the space ProDOS reserves for the
selector routine, so this crashes or hangs the system. Oops!

Reorder some of the code so this doesn't happen.