Launcher: Check for /RAM, warn if not empty
Warn if /RAM is present and not empty, as modules will disconnect it
causing contents to be lost.

This is not a 100% perfect check as DESKTOP.SYSTEM is only invoked
when first launching DeskTop. If you were to e.g. invoke BASIC.SYSTEM
from within DeskTop then /RAM will be reconnected, and then once again
disconnected when you BYE back to the DeskTop. In that case there's
not an easy way to recover without a massive overhaul. This warning is
hopefully sufficient.