Image Preview: Support for Packed Super Hi-Res PNT files ($C0/$0001)
The implementation is ridiculously slow, as each byte is written to
the SHR screen via a callback that increments an absolute pointer.

There's also no feedback that anything is happening while the image
loads, because the palette table has been cleared to black. Ah well.

(Note that most IIgs images are in $C0/$002 "Apple Preferred Format"
so this really doesn't do a whole lot, but it was easy.)