Electric Duet Preview: Detect Mockingboard and use for playback
Rather than requiring the user to select Mockingboard playback by
hitting a key, detect the card and switch to that player
automatically. The user can still override. The routine to patch the
playback code with the slot address is really dumb but it was easy.

This also revealed what seems to be a glitch with Virtual ][ - if
running in accelerated mode, the start of playback happens too fast.
Worked around by toggling the speaker inaudibly before playback.

Future work: Detect The Cricket! and use that for playback.