DeskTop.system: Prevent crash copying to RAMCard
The DeskTop.system launcher had a latent issue where the last bit of
the code ended up overwritten by the copy buffer when copying to
RAMCard. An assert would have caught this but wasn't present.
Previously, we got lucky - it was the tail end of a proc that would
only happen if non-standard files were present to copy.

The change in e83915ec pushed more code into the danger zone, and so a
crash was inevitable. Oops! Fix by adjusting buffer start location and
adding an assert to catch this in the future.