Image Preview: Improve LZ4FH load experience
The code would load/decompress the hires file, then restore the 10K
overlay, then convert the file from hires to double-hires. This caused
a stutter during the load. Additionally, if another LZ4FH image was
previewed the 10K overlay would be restored again.

Instead, set a flag and only restore the 10K overlay when the DA is
exiting.