DeskTop: Ensure correct window is refreshed after Rename
When renaming icons, usually no window refresh is needed as the icon
can be updated in situ. But if the icon is in a window sorted by name
the window contents need to be refreshed since the order may change.
This was erronously updating the active window, even if selection
was in another window.

Address this by activating the window containing selection if this is
needed, and also scroll the selected icon into view as a convenience.