DeskTop: Fix corruption when exiting Shortcuts dialogs
Introduced in c2975b5 (back in March 2023); rearranging the shortcut
overlay from $9000-$A000 down to $5000-$6000 put at least one routine
into the "danger zone" so that when the shortcuts routine (1) closed
the file picker and (2) cleared updates but (3) the overlay was not
restored, the update clearing would call into bad code leading to
corruption or a crash.

Move the routine into a safe zone. Also, from an audit of the routines
invoked by ClearUpdates, drop in several more asserts. This is
definitely not a perfect solution.

Fixes #790.