Launcher: Compute machine flags and stash so modules don't need to
Modules would query several system properties to modify behavior: * Is
IIc, IIc+, IIgs, Laser 128, or Mac IIe Card? * Has a Mega II chip? *
Has Le Chat Mauve Eve card?

... and then store flags for these somewhere that lib/ files could
find them. That's a lot of redundant code, and wouldn't change on a
given system so extra work.

Instead, move this logic into the launcher and sacrifice a byte in
settings to store this. It will be overwritten each session but this
is more convenient and removes some complex dependencies. In the
future it might be necessary to optimize flag usage a bit or expand
into two bytes.

One thing that wasn't done is putting the "is this an Enhanced IIe or
later?" test into a flag; the main usage of this is for the loader
code (to draw MouseText progress bars). The loaders don't have a
convenient way to read settings, so this was left alone.