Message DA: Pad string with space to reduce artifacts
The display is updated with this logic:

* Draw the text at the new position, mostly covering what was there
* Clear the rect to the left of the string
* Clear the rect to the right of the string

Note that the display is not double-buffered. Because MGTK draws text
from the top to the bottom, the top right edge is drawn many cycles
before the right rect is cleared. This means if the text is scrolling
to the left, the previous "frame"'s top right is visible briefly
before it gets erased by the rect clearing. This is particularly
noticable with strings like "DESKTOP".

Mitigate this by always padding the string with a leading and trailing
space. Since that's drawn in conjunction with the rest of the string,
that clears a few pixels on either side at the same time, which
greatly reduces the time artifacts are visible. This isn't perfect -
if the speed is increased then the padding doesn't extend far enough.
But at high speeds everything is a blurry zoom anyway so this is good
enough.