DeskTop: Move GetFileEntryBlock buffer locations
Since GFEB is called during recursive file copy which uses $800 when
enumerating directories, avoid using that. Right now there is no
conflict because GFEB is only used when relinking, which only happens
for same-volume file relinking, which operates only on selected files
and doesn't require recursion. But be more robust for future use.