DeskTop: Fix issues with replacing files/folders during copy
During a file copy, if creating the destination file failed because it
already exists, DeskTop would mutate the destination file to match the
source. This was bogus in two cases:

* When replacing a regular file with a folder, the storage type would
  be bogus and the directory file wouldn't be populated correctly.

* When replacing a folder with anything, the storage type would remain
  unchanged and the contents would be corrupted.

Replace this by explicitly deleting the destination file if needed,
and creating from scratch. Since recursive deletion is non-trivial,
for now an error is shown if a folder is being replaced. That could be
revisited in the future, but it's better than it was!