Rename/Duplicate: Show "too long" error without dismissing dialog
Previously, the New Folder command's dialog would check that the
overall path length including the new name fit, and if not an alert
would be shown without closing the dialog, allowing the user to adjust
the name.

In contrast, the Rename and Duplicate commands' dialogs would show the
alert but then close the dialog, so the user would have to re-run the
command to try again.

Align the New Folder, Rename, and Duplicate dialog logic so that when
the user clicks OK the path length is checked.