DeskTop: Option to preserve upper/lowercase in filenames
When MouseDesk was written, the ProDOS filesystem had no support for
mixed-case file and volume names. Because SHOUTY.NAMES are ugly,
MouseDesk applied a heuristic to Make.Names.Nicer. And since 8-bit
Apple users keep the Caps Lock on most of the time [Citation needed],
when prompted to enter a new name for a file or disk, this was fine.
You'd type a SHOUTY.NAME and it would be shown as Less.Shouty. Nice!

Then GS/OS came along and some spare bits in the filesystem were
repurposed as "case bits" to make filenames case-preserving while
still being backwards compatible. (But not case-sensitive, thank
goodness, because that's user-hostile and breaks compatibility!)

Several decades later, modern DeskTop was updated to *show* files and
volume names using the GS/OS bits if present, and the heuristic
otherwise. But it would not *set* the bits when entering a file or
volume name, because they'd be saved as SHOUTY.NAMES because we all
keep Caps Lock on most of the time!

But for those users brave enough run with Caps Lock off, there's now
an option to preserve uppercase and lowercase as typed. This will
cause File > New Folder, File > Rename, File > Duplicate, Special >
Format Disk and Special > Erase Disk to write out the GS/OS case bits
as typed. Turn it on and Go.CrAzY!

(Caveat: AppleWorks files have their *own* way of storing case bits,
because that was invented before GS/OS. Modern DeskTop always shows
AppleWorks files with uppercase/lowercase as entered in AppleWorks,
and renaming or duplicating them in DeskTop updates their special
bits. This is because AppleWorks users are very picky.)

For #352