LETK: Remove clipping logic, ensure correct pen usage
The line edit control set up a clip rect to prevent text from bleeding
out of the control, but couldn't handle the control being offscreen.
This is doable, but math is hard. Because of this the Map DA --
currently the only movable window with a line edit control -- had
logic to prevent being moved offscreen.

Digging back through the history of the code, the clipping was added
when line edits could contain full paths, and clipping was preferred
over other fixes. That's no longer the case - all uses of the line
edit can fit the entire string. So... just remove the clipping code,
allow the Map DA to be positioned anywhere, and early-exit if the
containing window is obscured like in BTK.

Also noticed in the Map DA: Control+X would clear the field to black
instead of white. So ensure the pen is set when clearing.