New DA: Round Corners
This is a hack that just paints rounded corners on the screen. They
don't persist if overdrawn. On the original Mac OS this was probably
done by having the QuickDraw region used for screen drawing shaped as
a rounded rect, which would naturally clip drawing. MGTK doesn't have
anything like that so it's just a hack.

Cute but useless, so it lives in Extras/