Replace uses of lib/muldiv.s with lib/muldiv16.s
lib/muldiv.s provided separate Multiply_16_8_16 and Divide_16_8_16
routines with 16- and 8-bit inputs and 16-bit outputs (and 8-bit
remainder). The inputs and outputs are A,X and Y.

lib/muldiv16.s provides a single MulDiv routine with 16-bit inputs and
16-bit outputs (including remainder). The inputs and outputs are zero
page locations.

The MulDiv routine is smaller, more general purpose, and handles a
larger range of values, although call setup is more elaborate. MulDiv
is primarily used for scaling values (b = a * n / d) but with a
numerator or divisor of 1 it can be used for either multiplication or
division.

MulDiv is already used in several places, so to simplify the code
overall, drop lib/muldiv.s. As a bridge, a few places define helpers
with the old signatures.

A follow-on commit will rename muldiv16 to muldiv.