DeskTop: Fix initial window size for list views
Investigating #792 turned up a glitch in computing the bounding box
for list views. Rather than calculating the actual width of all lines,
a fixed width is used to ensure long dates fit. But this was applied
(1) after the bounding box was inflated and (2) by assigning to a
right edge rather than as a width.

This ensures that when a window is opened in list view (which happens
if the parent window is a list view) then the new window's horizontal
scrollbars will be inactive.