Fonts: Add latin1 fonts for System/Monaco
Maintaining separate fonts for the different language encodings is a
pain - updating a glyph could require updating different locations
across multiple files.

To ease this: for System and Monaco fonts, include a 256-glyph font
covering the ISO-8859-1 "Latin 1" code page (which is basically the
same as Unicode U+0000 through U+00FF), and a tool to generate the
different fonts for each language.

These 256-glyph fonts are not used by A2D itself and unknown if
they're compatible with MGTK. This is just develop-time tooling.

The only visible change is: the § glyph in the French build now looks
slightly different, aligned with the glyph shape in other languages.

(Yes, I'm conflating language, region, script, etc. Sorry!)