MGTK: Fix title bar stripes when using 8px font
The commit 258ecee tweaked the title bar appearance, but it only
accounted for the standard 9px tall system font, and that was pure
luck. With an 8px tall font the stripes were drawn differently to the
left and right of the "go away" box.

Fix this properly. This only affects custom builds that use a
different font.

There is a picture of the bad title bars in this comment, but you need
to look closely since there's a menu in the way:
https://github.com/a2stuff/a2d/issues/793#issuecomment-2079638597