Font Preview: Don't crash if no icon is selected
If the SHOW.FONT.FILE file is added to the Apple Menu, it can be used
on arbitrary files, just in case the heuristics fail. It would crash
if no icon was selected. Re-order the init logic so the name isn't
parsed unless the file is proven to be valid. There was already a test
case for this.