DeskTop: Fix File > Duplicate if parent window's icon is showing
When duplicating a file, the containing window is refreshed prior to
the rename prompt showing. If the window's icon is present, it would
become selected. But the call to select the newly created file
wouldn't clear selection first, leading to invalid selection spanning
two windows.

As a safe fix, clear selection prior to selecting the new icon. A
better fix would be to avoid selecting the parent icon in the first
place, to avoid extra work, but that's in a widely used proc.

Likely introduced when changing duplicate from a modal to modeless.