LETK: Fix caret glitch when using Ctrl+X in rename field
The Line Edit logic for handling Ctrl+X (Clear on the IIgs) didn't
handle redrawing properly for a centered edit field, like icon
renaming in DeskTop: it would mispaint the caret on the left edge of
the field.

Rearrange the code to use the more general clear-and-refresh logic.

Already present in the test plan.