DeskTop: Show disk erase confirmation as an alert that wraps
Early in 1.4 the disk erase confirmation prompts were expanded and
when composed with the computed disk identifier could be as long as
"Are you sure you want to erase the DOS 3.3 disk in slot 6 drive 1?"
(longer in German, of course). This no longer fit in the dialog.

Using similar logic to that of Disk Copy, compose the full string into
a buffer and show it using an alert. This has the advantage of showing
an icon and beeping, which is probably the right thing to do before
erasing a disk anyway.